♻️ local 配置纯函数拆到 sibling .config.ts + 去 extension guard 的 local 硬编码 #130
