+++ frontend/src/components/asset/LocalConfigSection.tsx
@@ -3,42 +3,11 @@ import { useTranslation } from "react-i18next";
 import { Input, Label, Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@opskat/ui";
 import { ListLocalShells } from "../../../wailsjs/go/local/Local";
 import type { localterm_svc } from "../../../wailsjs/go/models";
-import { formatLocalShellArgs, parseLocalShellArgs } from "@/lib/localShellArgs";
+import { formatLocalShellArgs } from "@/lib/localShellArgs";
 import type { AssetFormHandle, ConfigSectionProps } from "@/lib/assetTypes/formContract";
+import { buildLocalConfig, parseLocalConfig, LOCAL_DEFAULTS, type LocalFormState } from "./LocalConfigSection.config";
 
 type ShellInfo = localterm_svc.ShellInfo;
-
-export interface LocalFormState {
-  shell: string;
-  args: string;
-  cwd: string;
-}
-
-export const LOCAL_DEFAULTS: LocalFormState = { shell: "", args: "", cwd: "~" };
-
-/** 保存序列化:镜像旧 handleSubmit 的 local 分支(行为保持)。args 非法→抛(调用方 toast)。 */
-export function buildLocalConfig(state: LocalFormState): string {
-  const cfg: Record<string, unknown> = {};
-  if (state.shell) cfg.shell = state.shell;
-  const argList = parseLocalShellArgs(state.args);
-  if (argList.length) cfg.args = argList;
-  if (state.cwd) cfg.cwd = state.cwd;
-  return JSON.stringify(cfg);
-}
-
-/** 编辑态回填:镜像旧 loadLocalConfig。解析失败→默认值。 */
-export function parseLocalConfig(configJSON: string): LocalFormState {
-  try {
-    const cfg = JSON.parse(configJSON || "{}");
-    return {
-      shell: cfg.shell || "",
-      args: formatLocalShellArgs(cfg.args || []),
-      cwd: cfg.cwd || "~",
-    };
-  } catch {
-    return { ...LOCAL_DEFAULTS };
-  }
-}
 
 export const LocalConfigSection = forwardRef<AssetFormHandle, ConfigSectionProps>(function LocalConfigSection(
   { editAsset, onValidityChange },
